Crate sp_session_validator_management

Primitives for Partner Chain committee selection.

This crate implements shared types and traits used to implement Partner Chain committee rotation.

Structs

AuthoritySelectionInputs

The part of data for selection of authorities that comes from the main chain. It is unfiltered, so the selection algorithm should filter out invalid candidates.

MainChainScripts

Collection of all mainchain script info needed for committee selection

Enums

CommitteeMember

Type representing committee members, either permissioned or registered

InherentError

Error type used for failing calls of the Committee Selection inherent.

Constants

INHERENT_IDENTIFIER

Inherent identifier used by the Committee Selection pallet

Traits

SessionValidatorManagementApi

Runtime API declaration for Session Validator Management